Version 2.0
The second playable version of the game. The player now has letters which can be collected around the level. After the letters are collected, they can access a door, which will then show the Caesar-Cipher encrypted message, and all of the letters that they collected and in whatever order they were collected in. The player must enter in the correct word. The correct word is "Apple". In the upcoming versions, this will be improved drastically, and be made easier for the player through the use of hints.